updated maven project to gradle project
