+++ build.gradle
@@ -1,70 +1,24 @@
-plugins {
-    id "com.jfrog.bintray" version "1.7.1"
-    id "java"
-    id "maven-publish"
-}
-
-description = """logentries-android"""
-
-repositories {
-    maven { url "http://repo.maven.apache.org/maven2" }
-}
-
-dependencies {
-    compile group: 'com.google.android', name: 'android', version:'4.1.1.4'
-    testCompile group: 'junit', name: 'junit', version:'3.8.1'
-}
-
-def versionNumber = '4.0.0'
-
-publishing {
-    publications {
-        mavenJava(MavenPublication) {
-            from components.java
-            groupId 'com.logentries'
-            artifactId 'logentries-android'
-            version versionNumber
-        }
+buildscript {
+    repositories {
+        jcenter()
+        maven { url "https://plugins.gradle.org/m2/" }
+    }
+    dependencies {
+        classpath 'com.android.tools.build:gradle:2.2.0-beta2'
+        classpath 'com.github.dcendents:android-maven-gradle-plugin:1.4.1'
     }
 }
 
-Properties properties = new Properties()
-properties.load(project.rootProject.file('local.properties').newDataInputStream())
-
-def binTrayUserOrg = properties.getProperty("bintray.userorg")
-def binTrayUser = properties.getProperty("bintray.user")
-def binTrayKey = properties.getProperty("bintray.apikey")
-def siteUrl = properties.getProperty("siturl")
-def gitUrl = properties.getProperty("giturl")
-def issueUrl = properties.getProperty("issueurl")
-
-bintray {
-    user = binTrayUser
-    key = binTrayKey
-
-    publications = ['mavenJava'] //When uploading Maven-based publication files
-    dryRun = false //Whether to run this as dry-run, without deploying
-    publish = true //If version should be auto published after an upload
-    pkg {
-        repo = 'maven'
-        name = 'le-android'
-        userOrg = binTrayUserOrg
-        desc = 'fork of the logentries android library'
-        websiteUrl = siteUrl
-        issueTrackerUrl = issueUrl
-        vcsUrl = gitUrl
-        licenses = ['Apache-2.0']
-        publicDownloadNumbers = true
-        version {
-            name = versionNumber
-            desc = 'Logentries for android'
-        }
+allprojects {
+    repositories {
+        jcenter()
     }
+}
+
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
 
 task wrapper(type: Wrapper) {
     gradleVersion = '3.0'
-}
-
-task install(dependsOn: build) {
 }
