+++ lib/.gitignore
@@ -1,5 +1,5 @@
 
-# Created by https://www.gitignore.io/api/jetbrains,java,gradle,android,windows,osx
+# Created by https://www.gitignore.io/api/jetbrains,windows,osx,android
 
 ### JetBrains ###
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio and Webstorm
@@ -56,53 +56,6 @@ fabric.properties
 # *.ipr
 
 
-### Android ###
-# Built application files
-*.apk
-*.ap_
-
-# Files for the ART/Dalvik VM
-*.dex
-
-# Java class files
-*.class
-
-# Generated files
-bin/
-gen/
-out/
-
-# Gradle files
-.gradle/
-build/
-
-# Local configuration file (sdk path, etc)
-local.properties
-
-# Proguard folder generated by Eclipse
-proguard/
-
-# Log Files
-*.log
-
-# Android Studio Navigation editor temp files
-.navigation/
-
-# Android Studio captures folder
-captures/
-
-# Intellij
-*.iml
-.idea/workspace.xml
-.idea/libraries
-
-# Keystore files
-*.jks
-
-### Android Patch ###
-gen-external-apklibs
-
-
 ### Windows ###
 # Windows image file caches
 Thumbs.db
@@ -153,33 +106,48 @@ Temporary Items
 .apdisk
 
 
-### Java ###
+### Android ###
+# Built application files
+*.apk
+*.ap_
+
+# Files for the ART/Dalvik VM
+*.dex
+
+# Java class files
 *.class
 
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
+# Generated files
+bin/
+gen/
+out/
 
-# Package Files #
-*.jar
-*.war
-*.ear
+# Gradle files
+.gradle/
+build/
 
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
+# Local configuration file (sdk path, etc)
+local.properties
 
+# Proguard folder generated by Eclipse
+proguard/
 
-### Gradle ###
-.gradle
-/build/
+# Log Files
+*.log
 
-# Ignore Gradle GUI config
-gradle-app.setting
+# Android Studio Navigation editor temp files
+.navigation/
 
-# Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
-!gradle-wrapper.jar
+# Android Studio captures folder
+captures/
 
-# Cache of project
-.gradletasknamecache
+# Intellij
+*.iml
+.idea/workspace.xml
+.idea/libraries
 
-# # Work around https://youtrack.jetbrains.com/issue/IDEA-116898
-# gradle/wrapper/gradle-wrapper.properties
+# Keystore files
+*.jks
+
+### Android Patch ###
+gen-external-apklibs
