--- local.properties
@@ -1,12 +1,0 @@
-## This file is automatically generated by Android Studio.
-# Do not modify this file -- YOUR CHANGES WILL BE ERASED!
-#
-# This file must *NOT* be checked into Version Control Systems,
-# as it contains information specific to your local configuration.
-#
-# Location of the SDK. This is only used by Gradle.
-# For customization when using a Version Control System, please read the
-# header note.
-#Fri Aug 26 08:01:27 CEST 2016
-ndk.dir=E\:\\android-sdk\\ndk-bundle
-sdk.dir=E\:\\android-sdk
